Added Ioannis' additional tests.
